requirements.txt
django-rest-swagger==2.0.7
djangorestframework==3.5.1